server/Python/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
openai==1.66.3
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
boto3==1.38.6
gunicorn==20.1.0
deep-translator==1.11.4
